% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_takuzu.R
\name{create_game}
\alias{create_game}
\title{Creer une grille}
\usage{
create_game()
}
\value{
une grille de taille 8x8
}
\description{
Cette fonction cree une grille de taille 8x8 qui respecte les 3 regles du jeu
}
\examples{
create_game()
}
