% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_tazuku.R
\name{create_game1}
\alias{create_game1}
\title{Creer une grille}
\usage{
create_game1()
}
\value{
une grille de taille 8x8 ou un bouleen FALSE
}
\description{
Cette fonction cree une grille de taille 8x8 respectant les regles : pas plus de 3 zeros ou 3 uns d'affiles et exactement 4 zeros et 4 uns pour chaque ligne et chaque colonne. Si a un moment donne la configuration de la grille ne permet pas de respecter ces regles, le code s'arrete avec un FALSE. Sinon, elle renvoie la grille complete
}
\examples{
create_game1()
}
