% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_tazuku.R
\name{create_game3}
\alias{create_game3}
\title{Creer une grille}
\usage{
create_game3()
}
\value{
une grille de taille 8x8
}
\description{
Cette fonction renvoie une grille de taille 8x8 respectant les 2 premières regles. Tant que la fonction create_game1 renvoie False, on continue jusqu'a obtenir une grille correcte
}
\examples{
create_game3()
}
