% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_takuzu.R
\name{create_line2}
\alias{create_line2}
\title{creer une ligne}
\usage{
create_line2(line)
}
\arguments{
\item{line}{un vecteur de taille 8 contenant des nombres 0 et 1 et des NaN}
}
\value{
un vecteur de taille 8 et un booleen
}
\description{
Cette fonction remplie un vecteur de taille 8 contenant des 0,des 1 et des NaN. Elle cherche a remplacer les NaN par des 0 ou 1 qui respectent les regles. Si la configuration de line ne permet pas le respect de celles-ci, elle renvoie FALSE. Sinon, elle renvoie le vecteur remplie avec un TRUE
}
\examples{
create_line2(c(1,1,0,0,NaN,NaN,NaN,NaN))
}
