% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_takuzu.R
\name{create_line3}
\alias{create_line3}
\title{Creer une ligne}
\usage{
create_line3(line)
}
\arguments{
\item{line}{un vecteur de taille 8 contenant des 0, des 1 et des NaN}
}
\description{
Cette fonction remplie un vecteur de taille 8 contenant des 0, des 1 et des NaN. Tant que la fonction create_line2 renvoie FALSE, on la refait jusqu'a avoir une ligne correcte
}
\examples{
create_line3(c(1,1,0,0,NaN,NaN,NaN,NaN))
}
