% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_tazuku.R
\name{data_base}
\alias{data_base}
\title{creer une base de donnees}
\usage{
data_base()
}
\value{
3 vecteurs contenant des grilles de taille 8x8
examples
data_base()
}
\description{
Catte fonction cree 3 listes contenant chacune 20 grille de jeu Takuzu de difficulte facile, moyen et difficile
}
