% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_difficile.R, R/DB_moyen.R
\docType{data}
\name{DB_moyen}
\alias{DB_moyen}
\title{Base de données difficile}
\usage{
data(DB_difficile)

data(DB_moyen)
}
\description{
Un jeu de données difficile pour Takuzu

Un jeu de données moyen pour Takuzu
}
\keyword{datasets}
