% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_takuzu.R
\name{unchained}
\alias{unchained}
\title{Enlever les chaines de caracteres}
\usage{
unchained(V)
}
\arguments{
\item{V}{vecteur de chaines de caracteres}
}
\description{
Cette fonction prend en argument un vecteur contenant des chaines de caracteres et remplace chaque chaine contenant un entier en cette entier sans la chaine
}
\examples{
unchained(c("1","1","0","0","NaN","NaN","NaN","NaN"))
}
