% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_tazuku.R
\name{verif_game}
\alias{verif_game}
\title{Verifier une grille}
\usage{
verif_game(M)
}
\arguments{
\item{M}{une grille de taille 8x8 contenant des 0 et 1}
}
\value{
un Booleen
}
\description{
Cette fonction verifie si une grille de taille 8x8 repectent la troisieme regle : pas de lignes identiques et pas de colonnes identiques
}
