% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet_tazuku.R
\name{verif_line}
\alias{verif_line}
\title{Verification de ligne}
\usage{
verif_line(l)
}
\arguments{
\item{l}{un vecteur de taille 8}
}
\value{
un booleen
}
\description{
Cette fonction verifie si une ligne ou colonne respecte les regles : pas plus de 3 zeros ou 3 uns d'affiles et exactement 4 zeros et 4 uns
}
\examples{
verif_line(c(1,1,0,0,1,1,0,0))
}
